feat(cli): inventory every MCP server that has run on this machine
