+++ Makefile
@@ -17,7 +17,7 @@ STATICCHECK_VERSION ?= v0.8.0
 # zone-specific.
 CHECK_TZ ?= UTC
 
-.PHONY: all build test vet staticcheck fmt fmt-check lint check clean
+.PHONY: all build test vet vet-cross staticcheck fmt fmt-check lint check clean
 
 all: check build
 
@@ -29,6 +29,24 @@ test:
 
 vet:
 	$(GO) vet $(PKG)
+
+# The platforms CI compiles on. The release ships Windows and macOS binaries and
+# the CI matrix vets Windows, but a local `go vet` only ever sees the host, so
+# platform-specific code is compiled here for the first time on somebody else's
+# pull request.
+CROSS_GOOS ?= windows darwin linux
+
+# vet-cross typechecks every package *and its tests* for each of those. Building
+# the binaries is not enough and was not: `GOOS=windows go build ./...` skips
+# test files entirely, so a test calling syscall.Mkfifo behind a runtime
+# GOOS check passed every local gate and broke the Windows job, because a
+# runtime skip does not stop the compiler needing the symbol. Guard that code
+# with a build tag, the way console_unix.go does, and this target proves it.
+vet-cross:
+	@for os in $(CROSS_GOOS); do \
+		echo "GOOS=$$os go vet $(PKG)"; \
+		GOOS=$$os $(GO) vet $(PKG) || exit 1; \
+	done
 
 # staticcheck catches non-idiomatic code (e.g. interface{} over any, dead code).
 # Run through `go run` at the pinned version rather than whatever binary is on
@@ -43,7 +61,7 @@ fmt:
 fmt-check:
 	@out="$$(gofmt -s -l .)"; if [ -n "$$out" ]; then echo "gofmt -s needed:"; echo "$$out"; exit 1; fi
 
-lint: vet staticcheck
+lint: vet vet-cross staticcheck
 
 # check is the pre-commit/CI gate, formatting, static analysis, and the full
 # test suite under the race detector (matching CI).
