+++ cmd/mcpsnoop/inventory_test.go
@@ -8,9 +8,7 @@ import (
 	"io"
 	"os"
 	"path/filepath"
-	"runtime"
 	"strings"
-	"syscall"
 	"testing"
 	"time"
 
@@ -637,45 +635,6 @@ func TestInventoryDistinguishesArgumentBoundaries(t *testing.T) {
 	_, stdout, _ := executeInventory(t, nil)
 	if strings.Count(stdout, `--name "a b"`) != 1 || strings.Count(stdout, "--name a b\n") != 1 {
 		t.Fatalf("the two commands do not render distinguishably:\n%s", stdout)
-	}
-}
-
-// TestInventoryDoesNotHangOnAFifo covers a directory anything can write to. Open
-// on a fifo blocks until a writer appears, and the command prints nothing until
-// the whole walk is done, so one of them stopped the answer entirely.
-func TestInventoryDoesNotHangOnAFifo(t *testing.T) {
-	if runtime.GOOS == "windows" {
-		t.Skip("no fifos")
-	}
-	t.Setenv("MCPSNOOP_HOME", t.TempDir())
-	t0 := time.Date(2026, 8, 1, 12, 0, 0, 0, time.UTC)
-	writeSessionLog(t, "good.jsonl", stdioSession(t, "s1", "srv", t0, "/proj", []string{"node", "i.js"}, false)...)
-	fifo := filepath.Join(paths.SessionsDir(), "pipe.jsonl")
-	if err := syscall.Mkfifo(fifo, 0o600); err != nil {
-		t.Skipf("cannot make a fifo here: %v", err)
-	}
-
-	done := make(chan inventory, 1)
-	go func() {
-		inv, err := takeInventory(paths.SessionsDir(), false)
-		if err == nil {
-			done <- inv
-		}
-		close(done)
-	}()
-	select {
-	case inv, ok := <-done:
-		if !ok {
-			t.Fatal("takeInventory failed on a directory holding a fifo")
-		}
-		if len(inv.Servers) != 1 {
-			t.Fatalf("servers = %d, want the one real log still reported", len(inv.Servers))
-		}
-		if inv.Skipped != 1 {
-			t.Fatalf("skipped = %d, want the fifo counted", inv.Skipped)
-		}
-	case <-time.After(5 * time.Second):
-		t.Fatal("a fifo in the sessions directory hung the whole command")
 	}
 }
